requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.14
botocore==1.35.14
click==8.1.7
fastapi==0.114.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.8
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
